feat: fetch pool info
