+++ src/main.rs
@@ -1,7 +1,6 @@
-mod simulator;
-
+mod get_pool;
+use get_pool::get_pool_data;
 use clap::Parser;
-use simulator::{fetch_historical_data, simulate_liquidity, calculate_position_fees};
 
 fn get_subgraph_url(chain: &str) -> &'static str {
     match chain {
@@ -17,33 +16,12 @@ fn get_subgraph_url(chain: &str) -> &'static str {
 
 #[derive(clap::Parser)]
 struct Opts {
-    /// The EVM chain that Uniswap V3 is deployed to
     #[clap(short, long)]
     chain: String,
 
     /// The Uniswap V3 token pair address
     #[clap(short, long)]
     pair: String,
-
-    /// The minimum price of the liquidity range
-    #[clap(short = 'm', long)]
-    min: f64,
-
-    /// The maximum price of the liquidity range
-    #[clap(short = 'n', long)]
-    max: f64,
-
-    /// The fee tier
-    #[clap(short, long)]
-    fee: f64,
-
-    /// The start time of the simulation
-    #[clap(short, long)]
-    start: u64,
-
-    /// The end time of the simulation
-    #[clap(short, long)]
-    end: u64,
 }
 
 fn main() {
@@ -51,17 +29,17 @@ fn main() {
 
     let chain = &opts.chain;
     let pair = &opts.pair;
-    let min = Some(opts.min);
-    let max = Some(opts.max);
-    let fee = opts.fee;
-    let start = opts.start;
-    let end = opts.end;
 
     let subgraph_url = get_subgraph_url(chain);
-    let historical_data = fetch_historical_data(subgraph_url, pair, start, end);
-    let simulation_results = simulate_liquidity(historical_data, min, max, fee);
+    let pool_data = get_pool_data(subgraph_url, pair);
 
-    for result in simulation_results {
-        println!("{:?}", result);
+    match pool_data {
+        Some(data) => {
+            // Process the data as needed
+            println!("Pool data: {:?}", data);
+        },
+        None => {
+            println!("No data found for pool {}", pair);
+        }
     }
 }
